Splits tracing and metrics OTLP providers into separate files (#1724)
* splits tracing and metrics otlp providers


* adds tests that cover the errors creating exporters


---------
