+++ pkg/telemetry/providers/otlp/tracing_test.go
@@ -17,7 +17,9 @@ func TestCreateTraceExporter(t *testing.T) {
 	tests := []struct {
 		name    string
 		config  Config
+		ctx     func() context.Context
 		wantErr bool
+		errMsg  string
 	}{
 		{
 			name: "valid config",
@@ -26,6 +28,7 @@ func TestCreateTraceExporter(t *testing.T) {
 				Headers:  map[string]string{"Authorization": "Bearer token"},
 				Insecure: true,
 			},
+			ctx:     func() context.Context { return context.Background() },
 			wantErr: false,
 		},
 		{
@@ -37,6 +40,7 @@ func TestCreateTraceExporter(t *testing.T) {
 					"x-env":     "test",
 				},
 			},
+			ctx:     func() context.Context { return context.Background() },
 			wantErr: false,
 		},
 		{
@@ -45,7 +49,22 @@ func TestCreateTraceExporter(t *testing.T) {
 				Endpoint: "secure.example.com:4318",
 				Insecure: false,
 			},
+			ctx:     func() context.Context { return context.Background() },
 			wantErr: false,
+		},
+		{
+			name: "error creating sdk-span-exporter due to error (cancelled context)",
+			config: Config{
+				Endpoint: "secure.example.com:4318",
+				Insecure: true,
+			},
+			ctx: func() context.Context {
+				ctx, cancel := context.WithCancel(context.Background())
+				cancel()
+				return ctx
+			},
+			wantErr: true,
+			errMsg:  "context canceled",
 		},
 	}
 
@@ -53,125 +72,20 @@ func TestCreateTraceExporter(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
 
-			ctx := context.Background()
+			ctx := tt.ctx()
 			exporter, err := createTraceExporter(ctx, tt.config)
 
 			if tt.wantErr {
 				assert.Error(t, err)
 				assert.Nil(t, exporter)
-			} else {
-				assert.NoError(t, err)
-				assert.NotNil(t, exporter)
-				// Clean up
-				_ = exporter.Shutdown(ctx)
-			}
-		})
-	}
-}
-
-func TestCreateMetricExporter(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name    string
-		config  Config
-		wantErr bool
-	}{
-		{
-			name: "valid config",
-			config: Config{
-				Endpoint: "localhost:4318",
-				Headers:  map[string]string{"x-api-key": "secret"},
-				Insecure: true,
-			},
-			wantErr: false,
-		},
-		{
-			name: "config without headers",
-			config: Config{
-				Endpoint: "localhost:4318",
-				Insecure: false,
-			},
-			wantErr: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			ctx := context.Background()
-			exporter, err := createMetricExporter(ctx, tt.config)
-
-			if tt.wantErr {
-				assert.Error(t, err)
-				assert.Nil(t, exporter)
-			} else {
-				assert.NoError(t, err)
-				assert.NotNil(t, exporter)
-				// Clean up
-				_ = exporter.Shutdown(ctx)
-			}
-		})
-	}
-}
-
-func TestNewMetricReader(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name    string
-		config  Config
-		wantErr bool
-		errMsg  string
-	}{
-		{
-			name: "valid config",
-			config: Config{
-				Endpoint: "localhost:4318",
-				Headers:  map[string]string{"Authorization": "Bearer token"},
-				Insecure: true,
-			},
-			wantErr: false,
-		},
-		{
-			name: "missing endpoint",
-			config: Config{
-				Headers: map[string]string{"Authorization": "Bearer token"},
-			},
-			wantErr: true,
-			errMsg:  "OTLP endpoint is required",
-		},
-		{
-			name: "config with custom headers",
-			config: Config{
-				Endpoint: "otel-collector.local:4318",
-				Headers: map[string]string{
-					"x-api-key": "secret",
-					"x-env":     "production",
-				},
-				Insecure: false,
-			},
-			wantErr: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			ctx := context.Background()
-			reader, err := NewMetricReader(ctx, tt.config)
-
-			if tt.wantErr {
-				assert.Error(t, err)
-				assert.Nil(t, reader)
 				if tt.errMsg != "" {
 					assert.Contains(t, err.Error(), tt.errMsg)
 				}
 			} else {
 				assert.NoError(t, err)
-				assert.NotNil(t, reader)
+				assert.NotNil(t, exporter)
+				// Clean up
+				_ = exporter.Shutdown(ctx)
 			}
 		})
 	}
@@ -184,7 +98,9 @@ func TestNewTracerProvider(t *testing.T) {
 		name       string
 		config     Config
 		expectNoOp bool
+		ctx        func() context.Context
 		wantErr    bool
+		errMsg     string
 	}{
 		{
 			name: "valid config with endpoint",
@@ -195,6 +111,7 @@ func TestNewTracerProvider(t *testing.T) {
 				Insecure:     true,
 			},
 			expectNoOp: false,
+			ctx:        func() context.Context { return context.Background() },
 			wantErr:    false,
 		},
 		{
@@ -203,6 +120,7 @@ func TestNewTracerProvider(t *testing.T) {
 				SamplingRate: 0.1,
 			},
 			expectNoOp: true,
+			ctx:        func() context.Context { return context.Background() },
 			wantErr:    false,
 		},
 		{
@@ -213,7 +131,24 @@ func TestNewTracerProvider(t *testing.T) {
 				Insecure:     true,
 			},
 			expectNoOp: false,
+			ctx:        func() context.Context { return context.Background() },
 			wantErr:    false,
+		},
+		{
+			name: "expect error creating trace exporter due to canceled context",
+			config: Config{
+				Endpoint:     "localhost:4318",
+				SamplingRate: 1.0, // Always sample
+				Insecure:     true,
+			},
+			expectNoOp: false,
+			ctx: func() context.Context {
+				ctx, cancel := context.WithCancel(context.Background())
+				cancel()
+				return ctx
+			},
+			wantErr: true,
+			errMsg:  "failed to create trace exporter",
 		},
 	}
 
@@ -221,7 +156,7 @@ func TestNewTracerProvider(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
 
-			ctx := context.Background()
+			ctx := tt.ctx()
 			res, err := resource.New(ctx,
 				resource.WithAttributes(
 					semconv.ServiceName("test-service"),
@@ -235,6 +170,9 @@ func TestNewTracerProvider(t *testing.T) {
 			if tt.wantErr {
 				assert.Error(t, err)
 				assert.Nil(t, provider)
+				if tt.errMsg != "" {
+					assert.Contains(t, err.Error(), tt.errMsg)
+				}
 			} else {
 				assert.NoError(t, err)
 				assert.NotNil(t, provider)
